Improvements for setting locale
The locale is currently set by the value from the HTML tag's `lang` attribute.
https://github.com/Itonomy/module-flowbox/blob/075ee836544f485e28e8743d3282c976202f4958/Block/Widget/Flow.php#L53

This includes the language, e.g. `en`, but not the region which makes it invalid for Flowbox.
https://help.getflowbox.com/en/articles/4485750-changing-the-language-region-of-your-flow#our-supported-locales.

The changes here fixes this to get the locale from the widget options,
or if a locale is not set in the widget options,
it will be gotten from the Magento store locale.
https://docs.magento.com/user-guide/configuration/general/general.html#locale-options

The reason for having an option in the widget, is that some Magento stores do not have their locale configured correctly.